feat(question): 增加试卷删除功能并优化相关逻辑
- 在 AssignmentMapper 中添加了根据 paperId 查询作业的方法
- 在 AssignmentService 中添加了检查试卷是否被使用的功能
- 在 AssignmentServiceImpl 中实现了检查试卷是否被使用的逻辑- 在 TeacherBusinessController 中添加了删除试卷的接口
- 在 TestPaperMapper 中添加了物理删除试卷的方法
- 在 TestPaperServiceImpl 中实现了逻辑删除和物理删除的逻辑

+++ src/main/java/com/example/service/question/impl/TestPaperServiceImpl.java
@@ -2,9 +2,11 @@ package com.example.service.question.impl;
 
 import com.example.mapper.question.TestPaperMapper;
 import com.example.model.question.TestPaper;
+import com.example.service.question.AssignmentService;
 import com.example.service.question.TestPaperService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -13,17 +15,27 @@ public class TestPaperServiceImpl implements TestPaperService {
 
     @Autowired
     private TestPaperMapper testPaperMapper;
+    @Autowired
+    private AssignmentService assignmentService;
 
     @Override
+    @Transactional
     public int insert(TestPaper testPaper) {
         return testPaperMapper.insert(testPaper);
     }
 
     @Override
+    @Transactional
     public int delete(Long id) {
-        return testPaperMapper.delete(id);
+        if (assignmentService.checkPaperUsed(id)){
+            return testPaperMapper.delete(id);
+        }
+        else {
+            return testPaperMapper.realDelete(id);
+        }
     }
 
+    @Transactional
     @Override
     public int update(TestPaper testPaper) {
         return testPaperMapper.update(testPaper);
